+++ lib/protocols/index.js
@@ -5,7 +5,7 @@
 
 module.exports = {
     drafts: require('./drafts')
-  , 7: require('./hybi-07-12')
-  , 8: require('./hybi-07-12')
-  , 13: require('./hybi-16')
+  , 7: require('./hybi')
+  , 8: require('./hybi')
+  , 13: require('./hybi')
 };
